% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_na.R
\name{ran_na}
\alias{ran_na}
\title{Randomly assign NA's for Certain Variables}
\usage{
ran_na(df, n, vars)
}
\arguments{
\item{df}{the dataframe that we are looking to assing NA's to}

\item{n}{the number of rows to assign NA's to or the proportion of observations to assign
NA's to given as a number < 1.}

\item{vars}{a character vector of variables that will be assigned NA}
}
\value{
a dataframe with NA values in the var columns for the observations that were sampled.
There will also be vars_save columns with the original values before assigning NA's
}
\description{
This function will assign NA's to variables specified for randomly sampled rows. It will
also save the old values before randomly assigning NA's
}
