% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_obs.R
\name{sample_obs}
\alias{sample_obs}
\title{Sample Observtions}
\usage{
sample_obs(df, n)
}
\arguments{
\item{df}{dataframe to sample}

\item{n}{integer or numeric < 1 and > 0}
}
\value{
a dataframe with number of observations specified by n, or the proportion of
observations specified by n.
}
\description{
This function will sample n observations, or if n < 1, it will sample the proportion
of the total obseravtion specified by n.
}
